+++ homework01/caesar.py
@@ -2,25 +2,7 @@
 This module provides functions to encrypt and decrypt text using a Caesar cipher.
 """
 
-
-def letter_number(letter: str) -> int:
-    """
-    This function takes a letter and returns its position in the alphabet.
-    """
-    if "A" <= letter <= "Z":
-        return ord(letter) - ord("A")
-    if "a" <= letter <= "z":
-        return ord(letter) - ord("a")
-    return -1
-
-
-def number_to_letter(number: int, is_upper: bool) -> str:
-    """
-    This function takes a number and returns the corresponding letter in the alphabet.
-    """
-    if is_upper:
-        return chr(number + ord("A"))
-    return chr(number + ord("a"))
+from cipher_utils import process_text
 
 
 def encrypt_caesar(plaintext: str, shift: int = 3) -> str:
@@ -35,16 +17,7 @@ def encrypt_caesar(plaintext: str, shift: int = 3) -> str:
     >>> encrypt_caesar("")
     ''
     """
-    ciphertext = ""
-    for letter in plaintext:
-        num = letter_number(letter)
-        if num != -1:
-            is_upper = letter.isupper()
-            shifted_num = (num + shift) % 26
-            ciphertext += number_to_letter(shifted_num, is_upper)
-        else:
-            ciphertext += letter
-    return ciphertext
+    return process_text(plaintext, [shift], encrypt=True)
 
 
 def decrypt_caesar(ciphertext: str, shift: int = 3) -> str:
@@ -59,13 +32,4 @@ def decrypt_caesar(ciphertext: str, shift: int = 3) -> str:
     >>> decrypt_caesar("")
     ''
     """
-    plaintext = ""
-    for letter in ciphertext:
-        num = letter_number(letter)
-        if num != -1:
-            is_upper = letter.isupper()
-            shifted_num = (num - shift) % 26
-            plaintext += number_to_letter(shifted_num, is_upper)
-        else:
-            plaintext += letter
-    return plaintext
+    return process_text(ciphertext, [shift], encrypt=False)
